+++ src/Ampersand/Misc/Installer.php
@@ -40,14 +40,21 @@ class Installer
 
         // TODO: add function to clear/delete current meta population
 
+        $this->cleanupMetaPopulation($model);
+
+        $model->getMetaPopulation()->import();
+
+        return $this;
+    }
+
+    public function cleanupMetaPopulation(Model $model): Installer
+    {
         // Cleanup interface atoms that are not defined (anymore) in the Ampersand model files
         foreach ($model->getConceptByLabel(ProtoContext::CPT_IFC)->getAllAtomObjects() as $ifcAtom) {
             if (!$model->interfaceExists($ifcAtom->getId())) {
                 $ifcAtom->delete();
             }
         }
-
-        $model->getMetaPopulation()->import();
 
         return $this;
     }